Change outdated 'github_request' API instructions for calling authenticated Octokit::Client.
Just updating the README to reflect current reality.

I banged my head against using github_request with errors after using the 'sinatra-auth-github-test' repository as a basis and trying to duplicate the code outside of it using the sinatra_auth_github gem from scratch. It took a while to realize that there was a possible version mismatch between projects.